refactoring interactive cli
